fix: check existing version before redownloading
This is a different way to prevent redownloading. With the prior approach, we kept each bin in its own version folder, but this meant you had to run tests with the same version override that you installed with. With this, you can use the version override forever based on the initial download.